move make_hash to utils
